+++ tests/testthat/test_bestNormalize.R
@@ -3,9 +3,9 @@ context('bestNormalize functionality')
 data(iris)
 train <- iris$Petal.Width
 
-BNobject <- suppressWarnings(bestNormalize(train, out_of_sample = FALSE))
-BNobject4 <- bestNormalize(train, allow_orderNorm = FALSE, out_of_sample = FALSE)
-BNobject5 <- suppressWarnings(bestNormalize(train, out_of_sample = TRUE))
+BNobject <- suppressWarnings(bestNormalize(train, out_of_sample = FALSE, quiet = T))
+BNobject4 <- bestNormalize(train, allow_orderNorm = FALSE, out_of_sample = FALSE, quiet = T)
+BNobject5 <- suppressWarnings(bestNormalize(train, out_of_sample = TRUE, quiet = T))
 
 # Test transformations
 test_that('BestNormalize transformations with positive data', {
@@ -16,9 +16,9 @@ test_that('BestNormalize transformations with positive data', {
 })
 
 # 
-BNobject <- suppressWarnings(bestNormalize(c(-1, train)))
-BNobject4 <- suppressWarnings(bestNormalize(c(-1, train), allow_orderNorm = FALSE, out_of_sample = FALSE))
-test_that('BestNormalize transformations with mixed data, out-of-sample', {
+BNobject <- suppressWarnings(bestNormalize(c(-1, train), quiet = T))
+BNobject4 <- suppressWarnings(bestNormalize(c(-1, train), allow_orderNorm = FALSE, out_of_sample = FALSE, quiet = T))
+test_that('BestNormalize transformations with mixed data, in-sample', {
   expect_equal(BNobject$x.t, predict(BNobject))
   expect_equal(BNobject$x, predict(BNobject, inverse = T))
   expect_equal(BNobject4$x.t, predict(BNobject4))
@@ -26,8 +26,8 @@ test_that('BestNormalize transformations with mixed data, out-of-sample', {
 })
 
 
-BNobject <- suppressWarnings(bestNormalize(c(-train), out_of_sample = FALSE))
-BNobject4 <- suppressWarnings(bestNormalize(c(-train), allow_orderNorm = FALSE, out_of_sample = FALSE))
+BNobject <- suppressWarnings(bestNormalize(c(-train), out_of_sample = FALSE, quiet = T))
+BNobject4 <- suppressWarnings(bestNormalize(c(-train), allow_orderNorm = FALSE, out_of_sample = FALSE, quiet = T))
 test_that('BestNormalize transformations with negative data', {
   expect_equal(BNobject$x.t, predict(BNobject))
   expect_equal(BNobject$x, predict(BNobject, inverse = T))
@@ -36,8 +36,8 @@ test_that('BestNormalize transformations with negative data', {
 })
 
 train2 <- c(train, -1, NA)
-BNobject <- suppressWarnings(bestNormalize(train2, out_of_sample = FALSE))
-BNobject4 <- suppressWarnings(bestNormalize(train2, allow_orderNorm = FALSE, out_of_sample = FALSE))
+BNobject <- suppressWarnings(bestNormalize(train2, out_of_sample = FALSE, quiet = T))
+BNobject4 <- suppressWarnings(bestNormalize(train2, allow_orderNorm = FALSE, out_of_sample = FALSE, quiet = T))
 test_that('BestNormalize transformations with mixed data and missing values', {
   expect_equal(BNobject$x.t, predict(BNobject))
   expect_equal(BNobject$x, predict(BNobject, inverse = T))
@@ -46,22 +46,23 @@ test_that('BestNormalize transformations with mixed data and missing values', {
 })
 
 test_that('bestNormalize handles missing original data', {
-  suppressWarnings(b <- bestNormalize(c(NA, train), out_of_sample = FALSE))
+  suppressWarnings(b <- bestNormalize(c(NA, train), out_of_sample = FALSE, quiet = T))
   expect_equal(as.numeric(NA), b$x.t[1])
   expect_equal(as.numeric(NA), predict(b)[1])
   expect_equal(as.numeric(NA), predict(b, inverse = TRUE)[1])
 })
 
 test_that('bestNormalize handles missing new data', {
-  suppressWarnings(b <- bestNormalize(train, out_of_sample = FALSE))
+  suppressWarnings(b <- bestNormalize(train, out_of_sample = FALSE, quiet = T))
   expect_equal(as.numeric(NA), predict(b, newdata = c(1, NA))[2])
   expect_equal(as.numeric(NA), predict(b, newdata = c(1, NA), inverse = TRUE)[2])
 })
 
 # Test standardize = FALSE
 train2 <- c(train, -1, NA)
-BNobject <- suppressWarnings(bestNormalize(train2, standardize = FALSE, out_of_sample = FALSE))
-BNobject4 <- suppressWarnings(bestNormalize(train2, standardize = FALSE, allow_orderNorm = FALSE, out_of_sample = FALSE))
+BNobject <- suppressWarnings(bestNormalize(train2, standardize = FALSE, out_of_sample = FALSE, quiet = T))
+BNobject4 <- suppressWarnings(bestNormalize(train2, standardize = FALSE, 
+                                            allow_orderNorm = FALSE, out_of_sample = FALSE, quiet = T))
 test_that('BestNormalize transformations without standardization', {
   expect_equal(BNobject$x.t, predict(BNobject))
   expect_equal(BNobject$x, predict(BNobject, inverse = T))
@@ -70,14 +71,14 @@ test_that('BestNormalize transformations without standardization', {
 })
 
 test_that('bestNormalize without standardization handles missing original data', {
-  suppressWarnings(b <- bestNormalize(c(NA, train), standardize = FALSE, out_of_sample = FALSE))
+  suppressWarnings(b <- bestNormalize(c(NA, train), standardize = FALSE, out_of_sample = FALSE, quiet = T))
   expect_equal(as.numeric(NA), b$x.t[1])
   expect_equal(as.numeric(NA), predict(b)[1])
   expect_equal(as.numeric(NA), predict(b, inverse = TRUE)[1])
 })
 
 test_that('bestNormalize without standardization handles missing new data', {
-  suppressWarnings(b <- bestNormalize(train, standardize = FALSE, out_of_sample = FALSE))
+  suppressWarnings(b <- bestNormalize(train, standardize = FALSE, out_of_sample = FALSE, quiet = T))
   expect_equal(as.numeric(NA), predict(b, newdata = c(1, NA))[2])
   expect_equal(as.numeric(NA), predict(b, newdata = c(1, NA), inverse = TRUE)[2])
 })
@@ -85,15 +86,27 @@ test_that('bestNormalize without standardization handles missing new data', {
 ## Test lambert functionality in bestNormalize
 test_that("bestNormalize works with lambert of type s", {
   skip_on_cran()
-  b <-  suppressWarnings(bestNormalize(train, allow_lambert_s = TRUE))
+  b <-  suppressWarnings(bestNormalize(train, allow_lambert_s = TRUE, quiet = T))
   expect_true(!is.null(b$other_transforms$lambert_s))
   expect_true(is.null(b$other_transforms$lambert_h))
 })
 
 test_that("bestNormalize works with lambert of type h", {
   skip_on_cran()
-  b <-  suppressWarnings(bestNormalize(train, allow_lambert_h = TRUE))
+  b <-  suppressWarnings(bestNormalize(train, allow_lambert_h = TRUE, quiet = T))
   expect_true(is.null(b$other_transforms$lambert_s))
   expect_true(!is.null(b$other_transforms$lambert_h))
 })
 
+## Test progress bar functionality in bestNormalize
+test_that("bestNormalize progress bar works", {
+  expect_output(suppressWarnings(bestNormalize(train, quiet = F)))
+  expect_output(suppressWarnings(bestNormalize(train, quiet = F, loo = T)))
+})
+
+test_that("bestNormalize works with lambert of type h", {
+  skip_on_cran()
+  b <-  suppressWarnings(bestNormalize(train, allow_lambert_h = TRUE, quiet = T))
+  expect_true(is.null(b$other_transforms$lambert_s))
+  expect_true(!is.null(b$other_transforms$lambert_h))
+})