+++ tests/testthat/test_bn_parallel.R
@@ -8,24 +8,41 @@ train <- iris$Petal.Width[rep(1:nrow(iris), each = 3)]
 
 cl <- parallel::makeCluster(2)
 
-test_that("Parallel functionality works", {
-  b <- bestNormalize(train, warn = FALSE, cluster = cl, r = 2)
+test_that("Parallel functionality works for RCV", {
+  b <- bestNormalize(train, warn = FALSE, cluster = cl, r = 2, quiet = T)
   expect_true(is.null(b$other_transforms$lambert_s))
   expect_true(is.null(b$other_transforms$lambert_h))
   
-  b <- bestNormalize(train, warn = FALSE, cluster = cl, allow_lambert_s = TRUE, r = 2)
+  b <- bestNormalize(train, warn = FALSE, cluster = cl, allow_lambert_s = TRUE, r = 2, quiet = T)
   expect_true(!is.null(b$other_transforms$lambert_s))
   expect_true(is.null(b$other_transforms$lambert_h))
   
-  b <- bestNormalize(train, warn = FALSE, cluster = cl, allow_lambert_h = TRUE, r = 2)
+  b <- bestNormalize(train, warn = FALSE, cluster = cl, allow_lambert_h = TRUE, r = 2, quiet = T)
   expect_true(is.null(b$other_transforms$lambert_s))
   expect_true(!is.null(b$other_transforms$lambert_h))
   
-  b <- bestNormalize(train, warn = FALSE, cluster = cl, allow_lambert_h = TRUE, allow_lambert_s = TRUE, r = 2)
+  b <- bestNormalize(train, warn = FALSE, cluster = cl, allow_lambert_h = TRUE, allow_lambert_s = TRUE, r = 2, quiet = T)
   expect_true(!is.null(b$other_transforms$lambert_s))
   expect_true(!is.null(b$other_transforms$lambert_h))
   
 }) 
+
+train <- iris$Petal.Width
+
+test_that("Parallel functionality works for LOO", {
+  b <- bestNormalize(train, warn = FALSE, cluster = cl, loo = T, quiet = T)
+  b <- bestNormalize(train, warn = FALSE, cluster = cl, allow_lambert_h = TRUE, 
+                     allow_lambert_s = TRUE, loo = T, quiet = T)
+  expect_true(!is.null(b$other_transforms$lambert_s))
+  expect_true(!is.null(b$other_transforms$lambert_h))
+  
+}) 
+
+test_that("bestNormalize progress bar works in parallel", {
+  expect_output(suppressWarnings(bestNormalize(train, warn = FALSE, cluster = cl, quiet = F)))
+  expect_output(suppressWarnings(bestNormalize(train, warn = FALSE, cluster = cl, quiet = F, loo = T)))
+})
+
 
 parallel::stopCluster(cl)